+++ data/duke.txt
@@ -2,6 +2,6 @@ T | Y | read book
 D | Y | return book | 2020-01-01
 E | Y | project meeting | 2020-02-16
 T | Y | join sports club
-T | N | play netball
+T | Y | play netball
 D | Y | tut 1 | 2020-09-09
 E | N | competition | 2020-10-10
